Implementado cadastro de imagens

+++ src/main/java/com/api/turistae/dtos/DadosReviewDTO.java
@@ -16,7 +16,7 @@ public class DadosReviewDTO {
     // Atributos
     private long id;
     private String texto;
-    private byte nota;
+    private int nota;
 
     // Relacionamentos
     private TurismoDTO turismo;
